+++ generators/app/templates/aciiFX/src/lib/route/service.route.ts
@@ -1,0 +1,29 @@
+import express = require('express');
+
+const router = express.Router();
+
+/**
+ * @method service status route
+ * @param {string} "/" the route
+ * @param {express.Request} req requested fields
+ * @param {express.Response} res Result
+ * @author Flowtastisch
+ * @memberof Aciiverse
+ * @date 15.08.2024
+ */
+router.get('/', (req: express.Request, res: express.Response) => {
+    const   today   = new Date(),
+            options = {
+                day:    '2-digit' as "2-digit" | "numeric",
+                month:  '2-digit' as "2-digit" | "numeric",
+                year:   'numeric' as "2-digit" | "numeric",
+                hour:   '2-digit' as "2-digit" | "numeric",
+                minute: '2-digit' as "2-digit" | "numeric",
+                second: '2-digit' as "2-digit" | "numeric"
+            };
+    res.status(200).send({
+        message: `Service alive! Time: ${today.toLocaleDateString('de-DE', options)}`
+    });
+});
+
+module.exports = router;
