+++ generators/app/templates/demo/index.html
@@ -32,5 +32,9 @@
             <h2>Logout</h2>
             <button onclick="onLogout()">Logout</button>
         </div>
+        <div class="form">
+            <h2>Service Status</h2>
+            <button onclick="onNavToServiceStatus()">Open Service Status</button>
+        </div>
     </body>
 </html>
